+++ FrontEnd/workbook/src/pages/Main.js
@@ -1,4 +1,4 @@
-import HeaderSignin from '../components/header/header_signin';
+import HeaderSignin from '../components/header/header';
 
 export default function Main() {
     return (
